+++ models/Wishlist.js
@@ -19,6 +19,7 @@ class Wishlist {
       if (!username) throw new BadRequestError("Username is required");
 
       const userId = await getUserId(username);
+      if(userId === 0 || !userId) throw new BadRequestError("User does not exist");
 
       // Get all products in wishlist using user id
       const getWishListQuery = `SELECT 
@@ -68,9 +69,10 @@ class Wishlist {
       if (!username) throw new BadRequestError("Username is required");
       if (!productId) throw new BadRequestError("Product ID is required");
       const userId = await getUserId(username);
+      if(userId === 0 || !userId) throw new BadRequestError("User does not exist");
 
       // add product to wishlist using user id and product id
-      const queryStatement = `INSERT INTO wishlist(user_id, product_id) VALUES($1, $2) RETURNING user_id, product_id`;
+      const queryStatement = `INSERT INTO wishlist(user_id, product_id) VALUES($1, $2) RETURNING user_id AS "userId", product_id AS "productId"`;
       const values = [userId, productId];
       const wishlistValues = await db.query(queryStatement, values);
 
@@ -92,7 +94,9 @@ class Wishlist {
     try {
       if (!username) throw new BadRequestError("Username is required");
       if (!productId) throw new BadRequestError("Product ID is required");
+      
       const userId = await getUserId(username);
+      if(userId === 0 || !userId) throw new BadRequestError("User does not exist");
       
       // Delete product from wishlist using user id and product id
       const queryStatement = `DELETE FROM wishlist 
@@ -119,7 +123,9 @@ class Wishlist {
    */
   static async removeAll(username) {
     try {
+      if (!username) throw new BadRequestError("Username is required");
       const userId = await getUserId(username);
+      if(userId === 0 || !userId) throw new BadRequestError("User does not exist");
       
       // Clear entire wishlist using user id
       const queryStatement = `DELETE FROM wishlist WHERE user_id = $1`;
